Move repetition into its own file.
